cmake-1.1 PG and port:qt5-kde-devel : complete CMAKE_MKSPEC implementation
Qt's cmake modules use CMAKE_MKSPEC to determine the location of the mkspec
headerfiles but hardcode this information when Qt is being built. This means
that CMake-based projects use potentially wrong headers when building with
another compiler than the one used when Qt was built.
This change makes CMAKE_MKSPEC an optional runtime variable with a hardcoded
default.
The cmake 1.1 PG uses the new mechanism to pass the qt_qmake_spec variable
from the Qt5 PG to CMake.

Committed from host : Bola