port:kf5-kdevelop-devel* : change shutdown patch
The DUChain shutdown hasn't hung for me in a while, the hanging I did
get was during global destruction and possibly related to the singleshot
QTimer created by the previous version of this patch. Replacing it with
a simple SIGALRM-based self-killer installed before we exit from the core
cleanup routine.

Committed from host : Bola