port:copperspice : construct convenience plugin directories
CS seems to insist of bundling everything with each application. We can
get around bundling the libraries quite easily, but not so much for
the plugins. Make interface endpoints for the plugin directories in an
appropriate location, applications can then be "bundled"with symlinks to
those interface dirs rather than duplicate the plugin(s).

Committed from host : Bola