port:samba4 : tested on Linux and add python variants
Note that there are rpath issues on Linux that I currently do not
understand (and don't exist in 4.8.9). Running smbclient or apps linking
to libsmbclient raises errors about missing symbols in (older) system versions
of libtdb.so.1 and liblzma.so.5. Despite the fact that ldd shows the
failing binaries link to the proper library under $prefix.

Committed from host : Bola