devport PG: dependency on the main port should be runtime
This is a bit debatable but for the purpose of the devport, the main
port's presence only becomes required when you try to use the devport
for its intended purpose, building a dependent.
Using depends_lib dependency creates a circular dependency on the main
port when doing the very first destroot phase of the devport, when the
main port isn't installed yet.

Committed from host : Bola