port:zfs : build the actual 2.1.6 code! port:zfs-devel : idem !!!
The previous 2.1.6.2 version was actually some unknown development version
probably preceding 2.1.6 . There is a 2.1.7.rc3 version too which I
got to build but it lacks a number of module (sysctl) parameters like
zfs_arc.meta_limit so I preferred not to provide it.
As a result, port:zfs-devel is identical to port:zfs except for the fact
that the zfs.kext is also built via MacPorts (and thus not signed).

Committed from host : Portia.local