LTO PG : initial test implementation of a +fromHost variant for Linux
This will go into libexec/macports/lib/macports1.0/macports.tcl when better tested
and ready for deployment.

The idea is to give every port on Linux a variant that makes it a simple placeholder
for the corresponding distro package(s). That way, no more need to drop dependencies
on Linux.
It will be the user's entire responsibility to ensure that the distro packages are
and remain all installed, but at least there will then be a way to assess which
ports are "not really" installed.

Committed from host : Bola