port:qt53-kde : (hopefully) fix the missing all-inclusivie source tarball
Qt have retired a lot of the Qt5 releases from their archive server, leaving
me no other choice to fetch the code from their git server. This change
 does just that if no source tarball is found in the distfiles/qt5 subdir.
The skeleton umbrella source tree is checked out at the desired version tag;
Qt's own init-repository script is patched to checkout of the corresponding
versions in the individual submodules and then invoked; the tree is pruned
(removal of the .git dirs and components not available in the current version)
and finally the entire directory is tarballed in the file that we used to be
able to download and its checksum saved to a file from which it can be sourced
later.
People who already had a previously downloaded tarball will be invited to create
the checksum file manually.

Committed from host : Bola