port:MacPorts* : more preparations for a major upgrade.
Makes port:MacPorts a stub, and introduces MacPorts-altdevel . That port
is for testing new releases in an alternative prefix, but should also
allow installing the port like any other port.
When built from a MP install in /opt/local (or /alt/local) it will install
into /alt/local (or /opt/local !). This means one should be able to install
it from the main install (in /opt/local), test the install in /alt/local,
and then install the same version of MacPorts-altlocal from /alt/local ...
upgrading the main install.
This is completely un-tested as of now.

Committed from host : Bola