port:MacPorts-pextlib : upgrade to 2.10.0/2.10.0.8 and overhaul
The non-devel port now alo has the patches required to build on Linux
and the Linux build depends on (the new) port:libbsd. Since this
requires port:libmd we now use that port as the MD5 provider, which
means there is no longer a build conflict with port:openssl-dev.

Committed from host : Bola