port:curl-ca-bundle : fail with a constructive error message
There are apparently Linux systems that ship an archaic or crippled
version of the unzip command which doesn't support the -a option.
I don't want to impose a fetch dependency on everyone because of that,
so just warn users of such systems that they'll need to install
port:unzip and can then try again.

Closes: #113
Committed from host : Bola