patch-updater-1.0 PG : new Portgroup
This one is intended to be included ONLY when patchfiles have to be
updated or just refreshed during a port upgrade.
It works by hijacking the `port patch` phase, applying each patch in
succession and creating a commit named after each successful patch
in a dedicated git topic branch.

Committed from host : Bola