port:zstd : upgrading sync with mainstream
We stick to our approach of building with CMake if port:cmake has
already been installed, and can thus make port:zstdConfig.cmake a
stub port.

The principle remains the same as in upstream though:
> port install zstdConfig.cmake
will have the effect of installing cmake and will then cause
port:zstd to be installed with the +cmake variant.

Committed from host : Bola